CI: isort fixup on 1.1.x (#37004)

+++ pandas/_libs/khash.pxd
@@ -1,5 +1,6 @@
 from cpython.object cimport PyObject
-from numpy cimport int64_t, uint64_t, int32_t, uint32_t, float64_t
+from numpy cimport float64_t, int32_t, int64_t, uint32_t, uint64_t
+
 
 cdef extern from "khash_python.h":
     ctypedef uint32_t khint_t
